[bug] fix for Issue #42

+++ Kamaran/resources/views/edit_order.blade.php
@@ -137,7 +137,7 @@
                 var total = 1;
                 $('.calc').each(function () {
                     if ($(this).val() != '') {
-                        total *= parseInt($(this).val());
+                        total *= parseFloat($(this).val());
                     }
                 });
                 $('#total').html(total);
